ai_engine/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.1


